requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
drf-spectacular==0.14.0
